train/requirements.txt
autogluon==1.1.0
mlflow-skinny==2.13.2
sagemaker-mlflow==0.1.0
